% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{knitroCpp}
\alias{knitroCpp}
\title{Knitro C++ Wrapper}
\usage{
knitroCpp(fcts, startValues, num_equality_constraints,
  num_inequality_constraints, nnzJ, RjacIndexCons, RjacIndexVars, ub, lb,
  optionsFile)
}
\arguments{
\item{fcts}{is an R list of functions that includes the \code{objFun}, \code{objGrad}, \code{c}, and \code{jac}.}

\item{startValues}{is a vector of start values}

\item{num_equality_constraints}{is an integer with the number of equality constraints in \code{c}}

\item{num_inequality_constraints}{is an integer with the number of inequality constraints in \code{c}}

\item{nnzJ}{is an integer with the number of non-zero objects in the Jacobian}

\item{RjacIndexCons}{is a vector of length \code{nnzJ}. Each element contains the index of a
particular constraint (i.e. the index of a row in the jacobian).}

\item{RjacIndexVars}{is a vector of length \code{nnzJ}. Each element contains the index of a
particular variable (i.e. the index of a column in the jacobian).}

\item{ub}{a vector of upper bounds for each element in x0}

\item{lb}{a vector lower bounds for each element in x0}

\item{optionsFile}{the location of the options file}
}
\value{
A list with the vector that minimizes the objective function, the final function value, and Knitro's exit status
}
\description{
This function is the standard C++ wrapper around knitro. It defines the object
\code{KTR_new} and defines a callback function that is used to evaluate the objective
function, the constraints, and gradients. The only deviation from the standard C++
wrapper is to use \code{UserParam} to pass the original R functions on to the C++
callback function.
}
\seealso{
http://www.artelys.com/tools/knitro_doc/2_userGuide/gettingStarted/startCallableLibrary.html
}

