% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/knitro.R
\name{knitro_ms}
\alias{knitro_ms}
\title{Call the KNITRO C++ interface using multiple start values}
\usage{
knitro_ms(objFun, objGrad = NULL, c_equality = NULL, c_inequality = NULL,
  jac = NULL, jacIndexCons = NULL, jacIndexVars = NULL, x0 = NA,
  lb = NULL, ub = NULL, optionsFile = "options.opt")
}
\arguments{
\item{objFun}{is a scalar valued R function that returns the objective function}

\item{objGrad}{is a vector-valued R function with the gradient}

\item{c_equality}{is a vector-valued R function with equality constraints}

\item{c_inequality}{is a vector-valued R function with inequality constraints}

\item{jac}{is a vector with the content of the Jacobian (sparse)}

\item{jacIndexCons}{refers to each element of jac and contains the number
of the constraint it refers to. Indexing is C++ compatible, i.e. the first
constraint has index 0}

\item{jacIndexVars}{refers to each element of jac and contains the number
of the variable it refers to. Indexing is C++ compatible, i.e. the first
variable has index 0}

\item{x0}{is a matrix with starting values}

\item{lb}{is a vector of lower bounds}

\item{ub}{is a vector of upper bounds}

\item{optionsFile}{is the path and filename of the options file.
If it does not exist, the function will create it}
}
\value{
a list with the final estimates, the function value, and KNITRO's exit status
}
\description{
This is a multi start version of knitro(). Uses a matrix as startvalues where
each row corresponds to one set of startvalues to be used. This version of
multi-start gives the user more control over the start values than KNITRO's
built-in version of multi-start. If you want to use the built-in version of
multi-start instead, you can do so via the options file.
}

